# AnimalSalvationArmyShelters/AnimalSalvationArmyShelters/AnimalSalvationArmyShelters/Models/AdoptionContact.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace AnimalSalvationArmyShelters.Models
{
    public class AdoptionContact
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public string Address { get; set; }
        public string Email { get; set; }
        public string PhoneNo { get; set; }        
    }
}